requirements.txt
mcp[cli]
python-pptx
comtypes
